셰이더 추가

+++ Engine/Public/Engine_Struct.h
@@ -246,7 +246,7 @@ namespace Engine
 		wstring SkillName;
 		int Sp{};
 		int HitCount{};
-		float Damage{};
+		int Damage{};
 		wstring DamageInfo;
 		SKILL_TARGET Target;
 		SKILL_RANGE Range;
